refactor: 모델타입 분리
