+++ OpenMarket/OpenMarket/Model/Image.swift
@@ -1,0 +1,22 @@
+//
+//  Image.swift
+//  OpenMarket
+//
+//  Created by Eunsoo KIM on 2022/01/04.
+//
+
+import Foundation
+
+struct Image: Codable {
+  let id: Int
+  let url: String
+  let thumbnailURL: String
+  let succeed: Bool
+  let issuedAt: String
+
+  enum CodingKeys: String, CodingKey {
+    case id, url, succeed
+    case thumbnailURL = "thumbnail_url"
+    case issuedAt = "issued_at"
+  }
+}
